+++ runtime/servletsrc/com/exedio/cope/sampler/SamplerItemCache.java
@@ -136,7 +136,7 @@ final class SamplerItemCache extends Item
 
 			if(s instanceof IntegerField)
 			{
-				if(((Integer)sv.value).intValue() != 0)
+				if((Integer)sv.value != 0)
 					return false;
 			}
 			else
