+++ runtime/src/com/exedio/cope/DayField.java
@@ -185,7 +185,7 @@ public final class DayField extends FunctionField<Day>
 	Day get(final Row row)
 	{
 		final Object cell = row.get(getColumn());
-		return cell==null ? null : DayColumn.getDay(((Integer)cell).intValue());
+		return cell==null ? null : DayColumn.getDay((Integer)cell);
 	}
 
 	@Override
