+++ runtime/src/com/exedio/cope/PlusView.java
@@ -88,7 +88,7 @@ public final class PlusView<E extends Number> extends NumberView<E>
 			{
 				if(sourceValue==null)
 					return null;
-				result += ((Integer)sourceValue).intValue();
+				result += (Integer)sourceValue;
 			}
 			return (E)Integer.valueOf(result);
 		}
@@ -99,7 +99,7 @@ public final class PlusView<E extends Number> extends NumberView<E>
 			{
 				if(sourceValue==null)
 					return null;
-				result += ((Long)sourceValue).longValue();
+				result += (Long)sourceValue;
 			}
 			return (E)Long.valueOf(result);
 		}
@@ -110,7 +110,7 @@ public final class PlusView<E extends Number> extends NumberView<E>
 			{
 				if(sourceValue==null)
 					return null;
-				result += ((Double)sourceValue).doubleValue();
+				result += (Double)sourceValue;
 			}
 			return (E)Double.valueOf(result);
 		}
